requirements.txt
tensorflow==2.20.0
google-genai==0.7.0
streamlit==1.38.0
python-dotenv==1.0.1
Pillow==10.2.0
